// src/types/inventoryVoucher/GstBillTypes.d.ts
export interface GstBillTableProps {
  loading: boolean;
  handleShowDeleteDialog: (id: number) => void;
  showDeleteDialog: boolean;
  setShowDeleteDialog: Dispatch<SetStateAction<boolean>>;
  setTempDeleteId: Dispatch<SetStateAction<number | null>>;
  handleDeleteGst: () => void;
  deleteGstBillLoading: boolean;
  handleShowPrint: (row: GstBillTableData) => void;
  printLoading?: boolean;
  currentPage: number;
  setCurrentPage: Dispatch<SetStateAction<number>>;
  lastPage: number;
  perPage: number;
  onPerPageChange: (perPage: number) => void;
}
// ...
export interface GstBillTableData {
  Id: number;
  Sales_Id?: number;
  Sales_Date?: string;
  Sales_No?: string;
  Bill_Date?: string;
  Bill_No?: string;
  Party_Name?: string;
  Gross_Amt?: number | string;
  Total?: number | string;
  Cgst_Amt?: number | string;
  Sgst_Amt?: number | string;
  Igst_Amt?: number | string;
  Round_Amt?: number | string;
  Discount?: number | string;
}